- Moved the APIRootView into the common module under the name APIRootBaseView, root view can now subclass this view and implement the get_routes method. - Created further test cases.
